Application metadata cache test
